Add PGXN support
It's overkill to use Docker to build PGXN releases, but the consistency
is nice, and it gave me a chance to use Alpine Linux.

Basically pulls down the PGXN files, reads the version and repo from
them, then pulls down the relevant tagged release from their repo. The
PGXN file is added to the code, re-archived with the correct name and
exported to the package shared volume.

Lots of copy-paste, but I don't see a need for elegance here.