Fully automate packaging
This change permits Travis to build Citus for a PGXN release. The logic
is substantially simpler than other automated builds: Travis builds all
commits and attempts to upload them (no signing for PGXN releases). The
PGXN endpoint will return a 409 status if a version already exists, so
we detect that and exit cleanly. Other statuses are errors.