Switch to 'fancy' versioning
I'm not 100% sure whether a self-conflicting package is best in the
RPM world, but despite the warning everything seems to work well.

Merging this will release citus60_95 and citus60_96. The version will
be 6.0.1.citus-2 and existing Citus users will be auto-upgraded. This
and all subsequent 'fancy' version packages will conflict with one
another, ensuring two versions cannot be installed at the same time.