Get the packaging working
Though the (yum) package name is 'citus-ha', within the project, the
extension name is declared as 'citusha', resulting in files such as
'citusha.so', 'citusha--1.0.0--sql', etc. The spec file was looking
for 'citus-ha'-named files and not finding them. To clarify, I added
a 'extname' variable set to 'citusha' and started using that.

Additionally, 9.6 appears to have some fatal warnings and the version
number needed a release, so I disabled 9.6 and added -1 to the version.